+++ Nahyeon/src/main/java/com/depromeet/nahyeon/common/controller/HealthCheckController.java
@@ -1,0 +1,17 @@
+package com.depromeet.nahyeon.common.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "헬스 체크")
+@RestController
+public class HealthCheckController {
+
+	@GetMapping("/health_check.html")
+	public ResponseEntity<Void> healthCheck() {
+		return ResponseEntity.ok().build();
+	}
+}
