--- Nahyeon/src/main/java/com/depromeet/nahyeon/exception/CertificationCodeNotMatchedException.java
@@ -1,8 +1,0 @@
-package com.depromeet.nahyeon.exception;
-
-public class CertificationCodeNotMatchedException extends RuntimeException {
-
-	public CertificationCodeNotMatchedException() {
-		super("자격 증명에 실패하였습니다.");
-	}
-}
