--- Nahyeon/src/main/java/com/depromeet/nahyeon/exception/ResourceNotFoundException.java
@@ -1,12 +1,0 @@
-package com.depromeet.nahyeon.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-	public ResourceNotFoundException(String dataSource, long id) {
-		super(dataSource + "에서 ID " + id + "를 찾을 수 없습니다.");
-	}
-
-	public ResourceNotFoundException(String dataSource, String email) {
-		super(dataSource + "에서 ID " + email + "를 찾을 수 없습니다.");
-	}
-}
