--- Nahyeon/src/main/java/com/depromeet/nahyeon/model/UserStatus.java
@@ -1,5 +1,0 @@
-package com.depromeet.nahyeon.model;
-
-public enum UserStatus {
-	PENDING, INACTIVE, ACTIVE
-}
