--- Nahyeon/src/main/java/com/depromeet/nahyeon/model/dto/PostUpdateDto.java
@@ -1,19 +1,0 @@
-package com.depromeet.nahyeon.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class PostUpdateDto {
-
-	private final String content;
-
-	@Builder
-	public PostUpdateDto(
-		@JsonProperty("content") String content
-	) {
-		this.content = content;
-	}
-}
