--- Nahyeon/src/main/java/com/depromeet/nahyeon/model/dto/UserCreateDto.java
@@ -1,25 +1,0 @@
-package com.depromeet.nahyeon.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class UserCreateDto {
-
-	private final String email;
-	private final String nickname;
-	private final String address;
-
-	@Builder
-	public UserCreateDto(
-		@JsonProperty("email") String email,
-		@JsonProperty("nickname") String nickname,
-		@JsonProperty("address") String address
-	) {
-		this.email = email;
-		this.nickname = nickname;
-		this.address = address;
-	}
-}
