--- Nahyeon/src/main/java/com/depromeet/nahyeon/model/dto/UserResponse.java
@@ -1,17 +1,0 @@
-package com.depromeet.nahyeon.model.dto;
-
-import com.depromeet.nahyeon.model.UserStatus;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserResponse {
-
-	private Long id;
-	private String email;
-	private String nickname;
-	private UserStatus status;
-	private Long lastLoginAt;
-}
