+++ Nahyeon/src/main/java/com/depromeet/nahyeon/post/domain/PostCreate.java
@@ -1,0 +1,22 @@
+package com.depromeet.nahyeon.post.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class PostCreate {
+
+	private final long writerId;
+	private final String content;
+
+	@Builder
+	public PostCreate(
+		@JsonProperty("writerId") long writerId,
+		@JsonProperty("content") String content
+	) {
+		this.writerId = writerId;
+		this.content = content;
+	}
+}
