+++ Nahyeon/src/main/java/com/depromeet/nahyeon/post/infrastructure/PostRepository.java
@@ -1,0 +1,6 @@
+package com.depromeet.nahyeon.post.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
+}
