--- Nahyeon/src/main/java/com/depromeet/nahyeon/repository/PostRepository.java
@@ -1,6 +1,0 @@
-package com.depromeet.nahyeon.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
-}
