+++ Nahyeon/src/main/java/com/depromeet/nahyeon/user/controller/response/UserResponse.java
@@ -1,0 +1,17 @@
+package com.depromeet.nahyeon.user.controller.response;
+
+import com.depromeet.nahyeon.user.domain.UserStatus;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserResponse {
+
+	private Long id;
+	private String email;
+	private String nickname;
+	private UserStatus status;
+	private Long lastLoginAt;
+}
