+++ Nahyeon/src/main/java/com/depromeet/nahyeon/user/domain/UserStatus.java
@@ -1,0 +1,5 @@
+package com.depromeet.nahyeon.user.domain;
+
+public enum UserStatus {
+	PENDING, INACTIVE, ACTIVE
+}
