+++ Nahyeon/src/test/java/com/depromeet/nahyeon/controller/PostCreateControllerTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlGroup;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.depromeet.nahyeon.model.dto.PostCreateDto;
+import com.depromeet.nahyeon.post.domain.PostCreate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
@@ -36,7 +36,7 @@ class PostCreateControllerTest {
 	@Test
 	void 사용자는_게시물을_작성할_수_있다() throws Exception {
 		// given
-		PostCreateDto postCreateDto = PostCreateDto.builder()
+		PostCreate postCreateDto = PostCreate.builder()
 			.content("post content")
 			.writerId(1)
 			.build();
